2023.04.28
* fixes

+++ datalakePipelines/createFederatedDatabaseInstance.bash
@@ -4,7 +4,6 @@ dbName="${1}"
 source init.conf
 tenantName="federatedDB-${dbName//_/-}"
 fileName="fedDBinput$$.json"
-curlData="@"${fileName}
 
 # get a list of the all the pipelines to include in the federated instance
 list_pipelines.bash "${dbName}" | jq  > "${fileName}"
@@ -16,28 +15,17 @@ fi
 
 # delete any old instance of the same name
 atlas --profile ${profile} dataFederation delete ${tenantName} --force > /dev/null 2>&1
-# broken
-# atlas --profile ${profile} dataFederation create ${name} -f ${fileName} -o json
-# atlas --profile ${profile} dataFederation update ${name} -f ${fileName} -o json
-
-# make a federated DB using the json created above
-# curl --user "${MONGODB_ATLAS_PUBLIC_KEY}:${MONGODB_ATLAS_PRIVATE_KEY}" --digest \
-curl --user "${publicKey}:${privateKey}" --digest \
---header "Accept: application/vnd.atlas.2023-02-01+json" --header "Content-Type: application/json" \
---include \
---request POST "https://cloud.mongodb.com/api/atlas/v2/groups/${groupId}/dataFederation/?pretty=true" \
---data "$curlData"
+out=$( atlas --profile ${profile} dataFederation create ${tenantName} -f ${fileName} -o json )
 
 if [[ $? == 0 ]]
 then 
-    printf "\n\nFederated Database created\n"
-    atlas --profile ${profile} dataFederation describe ${tenantName} -o json | jq .hostnames
+    eval hn=$( printf "${out}" | jq .hostnames[0] )
+    uri="mongodb://${userName}:${passWord}@${hn}/?tls=true&authSource=admin"
+    printf "A Federated Database was created with connection string: \n\t\"${uri}\"\n"
+    [[ -e "${fileName}" ]] && rm "${fileName}"
+    exit 0
 else
-    printf "\n\n* * * error: Federated Database creation failed\n"
+    printf "\n\n* * * error: The Federated Database creation failed\n"
+    [[ -e "${fileName}" ]] && rm "${fileName}"
     exit 1
 fi
-
-rm "${fileName}"
-
-##--request PATCH "https://cloud.mongodb.com/api/atlas/v2/groups/${groupId}/dataFederation/${tenantName}?pretty=true" \
-exit 0
